Feat: BaseCollectionViewCell 생성. MediaCollectionViewCell에 적용.
